GUI Updates
These are some temporary patchworks to make sure the GUI works
correctly during the demo. I was having some issues pertaining to state
and passing objects around. As a temporary measure, I just went back to
creating new objects. This is only for a few files and will be updated
later.